requirements.txt
pip==21.3.1
setuptools==60.1.0
wheel==0.37.1
psutil==5.8.0
influxdb==5.3.1
pyserial==3.5